requirements.txt
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.7
pillow==10.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-telegram-bot==21.1.1
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.29
typing_extensions==4.11.0
urllib3==2.2.1
